Create T3GameEndInfoCommandHandler class
